feat: supported legacy tool use for Claude 3 models
