+++ README.md
@@ -15,9 +15,9 @@ The following models support `POST SERVER_URL/openai/deployments/MODEL_NAME/chat
 |anthropic.claude-v1|text-to-text|✅|✅|❌|✅|
 |anthropic.claude-v2|text-to-text|✅|✅|❌|✅|
 |anthropic.claude-v2:1|text-to-text|✅|✅|✅|✅|
-|anthropic.claude-3-sonnet-20240229-v1:0|text-to-text, image-to-text|❌|❌|❌|❌|
-|anthropic.claude-3-haiku-20240307-v1:0|text-to-text, image-to-text|❌|❌|❌|❌|
-|anthropic.claude-3-opus-20240229-v1:0|text-to-text, image-to-text|❌|❌|❌|❌|
+|anthropic.claude-3-sonnet-20240229-v1:0|text-to-text, image-to-text|❌|❌|✅|❌|
+|anthropic.claude-3-haiku-20240307-v1:0|text-to-text, image-to-text|❌|❌|✅|❌|
+|anthropic.claude-3-opus-20240229-v1:0|text-to-text, image-to-text|❌|❌|✅|❌|
 |stability.stable-diffusion-xl|text-to-image|❌|✅|❌|❌|
 |meta.llama2-13b-chat-v1|text-to-text|✅|✅|❌|❌|
 |meta.llama2-70b-chat-v1|text-to-text|✅|✅|❌|❌|
