+++ aidial_adapter_bedrock/llm/model/adapter.py
@@ -5,9 +5,11 @@ from aidial_adapter_bedrock.llm.bedrock_models import BedrockDeployment
 from aidial_adapter_bedrock.llm.chat_model import ChatCompletionAdapter
 from aidial_adapter_bedrock.llm.model.ai21 import AI21Adapter
 from aidial_adapter_bedrock.llm.model.amazon import AmazonAdapter
-from aidial_adapter_bedrock.llm.model.anthropic import (
-    AnthropicAdapter,
-    AnthropicChat,
+from aidial_adapter_bedrock.llm.model.claude.v1_v2.adapter import (
+    Adapter as Claude_V1_V2,
+)
+from aidial_adapter_bedrock.llm.model.claude.v3.adapter import (
+    Adapter as Claude_V3,
 )
 from aidial_adapter_bedrock.llm.model.cohere import CohereAdapter
 from aidial_adapter_bedrock.llm.model.meta import MetaAdapter
@@ -24,13 +26,13 @@ async def get_bedrock_adapter(
             | BedrockDeployment.ANTHROPIC_CLAUDE_V3_HAIKU
             | BedrockDeployment.ANTHROPIC_CLAUDE_V3_OPUS
         ):
-            return AnthropicChat.create(model, region, headers)
+            return Claude_V3.create(model, region, headers)
         case (
             BedrockDeployment.ANTHROPIC_CLAUDE_INSTANT_V1
             | BedrockDeployment.ANTHROPIC_CLAUDE_V2
             | BedrockDeployment.ANTHROPIC_CLAUDE_V2_1
         ):
-            return await AnthropicAdapter.create(
+            return await Claude_V1_V2.create(
                 await Bedrock.acreate(region), model
             )
         case (
