+++ aidial_adapter_bedrock/llm/model/cohere.py
@@ -1,5 +1,6 @@
 from typing import Any, AsyncIterator, Dict, List, Optional
 
+from aidial_sdk.chat_completion import Message
 from pydantic import BaseModel, Field
 from typing_extensions import override
 
@@ -18,7 +19,6 @@ from aidial_adapter_bedrock.llm.chat_model import (
     default_partitioner,
 )
 from aidial_adapter_bedrock.llm.consumer import Consumer
-from aidial_adapter_bedrock.llm.message import BaseMessage
 from aidial_adapter_bedrock.llm.model.conf import DEFAULT_MAX_TOKENS_COHERE
 from aidial_adapter_bedrock.llm.tokenize import default_tokenize_string
 from aidial_adapter_bedrock.llm.tools.default_emulator import (
@@ -153,10 +153,8 @@ class CohereAdapter(PseudoChatModel):
         )
 
     @override
-    def validate_base_messages(
-        self, messages: List[BaseMessage]
-    ) -> List[BaseMessage]:
-        messages = super().validate_base_messages(messages)
+    def preprocess_messages(self, messages: List[Message]) -> List[Message]:
+        messages = super().preprocess_messages(messages)
 
         # Cohere doesn't support empty messages,
         # so replace it with a single space.
