+++ aidial_adapter_bedrock/llm/model/meta.py
@@ -1,5 +1,6 @@
 from typing import Any, AsyncIterator, Dict, List, Optional
 
+from aidial_sdk.chat_completion import Message
 from typing_extensions import override
 
 from aidial_adapter_bedrock.bedrock import (
@@ -10,7 +11,6 @@ from aidial_adapter_bedrock.dial_api.request import ModelParameters
 from aidial_adapter_bedrock.dial_api.token_usage import TokenUsage
 from aidial_adapter_bedrock.llm.chat_model import PseudoChatModel
 from aidial_adapter_bedrock.llm.consumer import Consumer
-from aidial_adapter_bedrock.llm.message import BaseMessage
 from aidial_adapter_bedrock.llm.model.conf import DEFAULT_MAX_TOKENS_META
 from aidial_adapter_bedrock.llm.model.llama_chat import (
     llama_emulator,
@@ -93,10 +93,8 @@ class MetaAdapter(PseudoChatModel):
         )
 
     @override
-    def validate_base_messages(
-        self, messages: List[BaseMessage]
-    ) -> List[BaseMessage]:
-        messages = super().validate_base_messages(messages)
+    def preprocess_messages(self, messages: List[Message]) -> List[Message]:
+        messages = super().preprocess_messages(messages)
 
         # Llama behaves strangely on empty prompt:
         # it generate empty string, but claims to used up all available completion tokens.
