+++ aidial_adapter_bedrock/llm/tools/default_emulator.py
@@ -1,4 +1,4 @@
-from typing import List, Optional, Tuple
+from typing import List, Optional
 
 from aidial_adapter_bedrock.llm.message import (
     AIFunctionCallMessage,
@@ -14,17 +14,19 @@ from aidial_adapter_bedrock.utils.log_config import bedrock_logger as log
 class DefaultToolsEmulator(ToolsEmulator):
     def add_tool_declarations(
         self, messages: List[BaseMessage]
-    ) -> Tuple[List[BaseMessage], List[str]]:
+    ) -> List[BaseMessage]:
         if self.tool_config is not None:
             log.warning(
                 "The model doesn't support tools/functions, however they were "
                 "specified in the request. Continuing without tools/functions."
             )
-        return messages, []
+        return messages
+
+    def get_stop_sequences(self) -> List[str]:
+        return []
 
     def convert_to_base_messages(
-        self,
-        messages: List[BaseMessage | ToolMessage],
+        self, messages: List[BaseMessage | ToolMessage]
     ) -> List[BaseMessage]:
         ret: List[BaseMessage] = [
             msg for msg in messages if not isinstance(msg, ToolMessage)
